# improved pairwiseDisjoint performance